+++ adapters/risc0/src/guest.rs
@@ -1,6 +1,6 @@
 #[cfg(target_os = "zkvm")]
 use risc0_zkvm::guest::env;
-use sov_rollup_interface::zk::traits::{Zkvm, ZkvmGuest};
+use sov_rollup_interface::zk::{Zkvm, ZkvmGuest};
 
 use crate::Risc0MethodId;
