+++ adapters/risc0/src/host.rs
@@ -6,7 +6,7 @@ use risc0_zkvm::receipt::verify_with_hal;
 use risc0_zkvm::serde::to_vec;
 use risc0_zkvm::sha::Impl;
 use risc0_zkvm::{Prover, Receipt};
-use sov_rollup_interface::zk::traits::{Zkvm, ZkvmHost};
+use sov_rollup_interface::zk::{Zkvm, ZkvmHost};
 
 use crate::Risc0MethodId;
