+++ examples/demo-simple-stf/src/lib.rs
@@ -3,7 +3,7 @@ use std::io::Read;
 use sha2::Digest;
 use sov_rollup_interface::da::BlobTransactionTrait;
 use sov_rollup_interface::stf::{BatchReceipt, StateTransitionFunction};
-use sov_rollup_interface::zk::traits::Zkvm;
+use sov_rollup_interface::zk::Zkvm;
 
 #[derive(PartialEq, Debug, Clone, Eq, serde::Serialize, serde::Deserialize)]
