+++ examples/demo-simple-stf/tests/stf_test.rs
@@ -3,7 +3,7 @@ use std::fmt::Display;
 use demo_simple_stf::{ApplyBlobResult, CheckHashPreimageStf};
 use sov_rollup_interface::mocks::{MockZkvm, TestBlob};
 use sov_rollup_interface::stf::StateTransitionFunction;
-use sov_rollup_interface::traits::AddressTrait;
+use sov_rollup_interface::AddressTrait;
 
 #[derive(PartialEq, Debug, Clone, Eq, serde::Serialize, serde::Deserialize)]
 pub struct DaAddress {
