+++ module-system/module-implementations/sov-prover-incentives/src/call.rs
@@ -4,7 +4,7 @@ use anyhow::Result;
 use borsh::{BorshDeserialize, BorshSerialize};
 use sov_bank::Coins;
 use sov_modules_api::CallResponse;
-use sov_rollup_interface::zk::traits::Zkvm;
+use sov_rollup_interface::zk::Zkvm;
 use sov_state::WorkingSet;
 
 use crate::ProverIncentives;
