+++ module-system/module-implementations/sov-prover-incentives/src/query.rs
@@ -1,5 +1,5 @@
 use serde::{Deserialize, Serialize};
-use sov_rollup_interface::zk::traits::Zkvm;
+use sov_rollup_interface::zk::Zkvm;
 use sov_state::WorkingSet;
 
 use super::ProverIncentives;
