+++ module-system/sov-modules-api/src/default_context.rs
@@ -1,12 +1,13 @@
 #[cfg(feature = "native")]
 use serde::{Deserialize, Serialize};
 use sov_rollup_interface::crypto::SimpleHasher;
+use sov_rollup_interface::AddressTrait;
 #[cfg(feature = "native")]
 use sov_state::ProverStorage;
 use sov_state::{ArrayWitness, DefaultStorageSpec, ZkStorage};
 
 use crate::default_signature::{DefaultPublicKey, DefaultSignature};
-use crate::{Address, AddressTrait, Context, PublicKey, Spec};
+use crate::{Address, Context, PublicKey, Spec};
 
 #[cfg(feature = "native")]
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
