+++ module-system/sov-modules-api/src/lib.rs
@@ -23,7 +23,7 @@ pub use prefix::Prefix;
 pub use response::CallResponse;
 use serde::{Deserialize, Serialize};
 pub use sov_rollup_interface::crypto::SimpleHasher as Hasher;
-pub use sov_rollup_interface::traits::AddressTrait;
+pub use sov_rollup_interface::AddressTrait;
 use sov_state::{Storage, Witness, WorkingSet};
 use thiserror::Error;
