+++ rollup-interface/src/node/services/batch_builder.rs
@@ -1,3 +1,5 @@
+//! This module defines the trait that is used to build batches of transactions.
+
 /// BlockBuilder trait is responsible for managing mempool and building batches.
 pub trait BatchBuilder {
     /// Accept a new transaction.
