+++ rollup-interface/src/node/services/mod.rs
@@ -1,3 +1,7 @@
+//! The services module contains traits for the long-lived services which
+//! the full node uses to run the state transition function and serve user
+//! requests.
+
 pub mod batch_builder;
 pub mod da;
 pub mod stf_runner;
