+++ rollup-interface/src/state_machine/mocks.rs
@@ -1,3 +1,5 @@
+//! Defines mock instantiations of many important traits, which are useful
+//! for testing, fuzzing, and benchmarking.
 use std::fmt::Display;
 use std::io::Write;
 
@@ -7,11 +9,12 @@ use bytes::Bytes;
 use serde::{Deserialize, Serialize};
 use sha2::Digest;
 
-use crate::da::{BlobTransactionTrait, BlockHashTrait, CountedBufReader, DaSpec};
+use crate::da::{BlobTransactionTrait, BlockHashTrait, BlockHeaderTrait, CountedBufReader, DaSpec};
 use crate::services::da::SlotData;
-use crate::traits::{AddressTrait, BlockHeaderTrait, CanonicalHash};
-use crate::zk::traits::{Matches, Zkvm};
+use crate::zk::{Matches, Zkvm};
+use crate::AddressTrait;
 
+/// A mock commitment to a particular zkVM program.
 #[derive(Debug, Clone, PartialEq, Eq, BorshDeserialize, BorshSerialize, Serialize, Deserialize)]
 pub struct MockCodeCommitment(pub [u8; 32]);
 
@@ -21,14 +24,19 @@ impl Matches<MockCodeCommitment> for MockCodeCommitment {
     }
 }
 
+/// A mock proof generated by a zkVM.
 #[derive(Debug, Clone, PartialEq, Eq, BorshDeserialize, BorshSerialize, Serialize, Deserialize)]
 pub struct MockProof<'a> {
+    /// The ID of the program this proof might be valid for.
     pub program_id: MockCodeCommitment,
+    /// Whether the proof is valid.
     pub is_valid: bool,
+    /// The tamper-proof outputs of the proof.
     pub log: &'a [u8],
 }
 
 impl<'a> MockProof<'a> {
+    /// Serializes a proof into a writer.
     pub fn encode(&self, mut writer: impl Write) {
         writer.write_all(&self.program_id.0).unwrap();
         let is_valid_byte = if self.is_valid { 1 } else { 0 };
@@ -36,12 +44,14 @@ impl<'a> MockProof<'a> {
         writer.write_all(self.log).unwrap();
     }
 
+    /// Serializes a proof into a vector.
     pub fn encode_to_vec(&self) -> Vec<u8> {
         let mut encoded = Vec::new();
         self.encode(&mut encoded);
         encoded
     }
 
+    /// Tries to deserialize a proof from a byte slice.
     pub fn decode(input: &'a [u8]) -> Result<Self, anyhow::Error> {
         ensure!(input.len() >= 33, "Input is too short");
         let program_id = MockCodeCommitment(input[0..32].try_into().unwrap());
@@ -55,6 +65,7 @@ impl<'a> MockProof<'a> {
     }
 }
 
+/// A mock implementing the zkVM trait.
 pub struct MockZkvm;
 
 impl Zkvm for MockZkvm {
@@ -91,6 +102,7 @@ fn test_mock_proof_roundtrip() {
     assert_eq!(proof, decoded);
 }
 
+/// A mock address type used for testing. Internally, this type is standard 32 byte array.
 #[derive(Debug, PartialEq, Clone, Eq, serde::Serialize, serde::Deserialize)]
 pub struct MockAddress {
     addr: [u8; 32],
@@ -137,6 +149,8 @@ impl AddressTrait for MockAddress {}
     serde::Serialize,
     serde::Deserialize,
 )]
+
+/// A mock BlobTransaction from a DA layer used for testing.
 pub struct TestBlob<Address> {
     address: Address,
     hash: [u8; 32],
@@ -165,6 +179,7 @@ impl<Address: AddressTrait> BlobTransactionTrait for TestBlob<Address> {
 }
 
 impl<Address: AddressTrait> TestBlob<Address> {
+    /// Creates a new mock blob with the given data, claiming to have been published by the provided address.
     pub fn new(data: Vec<u8>, address: Address, hash: [u8; 32]) -> Self {
         Self {
             address,
@@ -174,6 +189,7 @@ impl<Address: AddressTrait> TestBlob<Address> {
     }
 }
 
+/// A mock hash digest.
 #[derive(Clone, Copy, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
 pub struct TestHash(pub [u8; 32]);
 
@@ -185,17 +201,11 @@ impl AsRef<[u8]> for TestHash {
 
 impl BlockHashTrait for TestHash {}
 
+/// A mock block header used for testing.
 #[derive(Serialize, Deserialize, PartialEq, core::fmt::Debug, Clone)]
 pub struct TestBlockHeader {
+    /// The hash of the previous block.
     pub prev_hash: TestHash,
-}
-
-impl CanonicalHash for TestBlockHeader {
-    type Output = TestHash;
-
-    fn hash(&self) -> Self::Output {
-        TestHash(sha2::Sha256::digest(self.prev_hash.0).into())
-    }
 }
 
 impl BlockHeaderTrait for TestBlockHeader {
@@ -204,12 +214,20 @@ impl BlockHeaderTrait for TestBlockHeader {
     fn prev_hash(&self) -> Self::Hash {
         self.prev_hash
     }
+
+    fn hash(&self) -> Self::Hash {
+        TestHash(sha2::Sha256::digest(self.prev_hash.0).into())
+    }
 }
 
+/// A mock block type used for testing.
 #[derive(Serialize, Deserialize, PartialEq, core::fmt::Debug, Clone)]
 pub struct TestBlock {
+    /// The hash of this block.
     pub curr_hash: [u8; 32],
+    /// The header of this block.
     pub header: TestBlockHeader,
+    /// The height of this block
     pub height: u64,
 }
 
@@ -224,6 +242,7 @@ impl SlotData for TestBlock {
     }
 }
 
+/// A [`DaSpec`] suitable for testing.
 pub struct MockDaSpec;
 
 impl DaSpec for MockDaSpec {
